Deprecate and move classes
