+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v2/debug/ChainHeadV2.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.api.response.v1.debug;
+package tech.pegasys.teku.api.response.v2.debug;
 
 import static tech.pegasys.teku.api.schema.SchemaConstants.DESCRIPTION_BYTES32;
