+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v2/debug/GetChainHeadsResponseV2.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.api.response.v1.debug;
+package tech.pegasys.teku.api.response.v2.debug;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
